--- src/features/project/ui/introduce-section/index.tsx
@@ -1,3 +1,0 @@
-export { SelectButton } from './SelectButton';
-export { TeamForm } from './TeamForm';
-export { StackForm } from './StackForm';
