--- src/features/project/ui/main-section/index.tsx
@@ -1,4 +1,0 @@
-export { MockupImageUploader } from './ImageUploadButton';
-export { AppIconUploader } from './ImageUploadButton';
-export { LinkInput, TitleInput, SubTitleInput } from './Inputs';
-export { ToggleActiveButton } from './ToggleActiveButton';
